refactor: param 조작 별도의 함수로 분리
