+++ src/shared/domain/addMoreMovies.ts
@@ -3,25 +3,20 @@ import { getSearchedPost } from "../../features/search/api/getSearchedPost";
 import { addMoviePost } from "../ui/renderers/addMoviePost";
 import { disableMoreButton } from "../ui/renderers/disabledMoreButton";
 import { showErrorPage } from "../ui/renderers/showErrorPage";
+import { getParams } from "./getParams";
+import { setParams } from "./setParams";
 
 export async function addMoreMovies($movieList: HTMLElement) {
-  const params = new URLSearchParams(window.location.search);
-  const query = params.get("query") ?? null;
-  const pageStr = params.get("page");
-  const currentPage = pageStr ? parseInt(pageStr) : 1;
-  const nextPage = currentPage + 1;
+  const { query, currentPage, nextPage } = getParams(
+    new URL(window.location.href)
+  );
+  setParams(query ?? "", nextPage);
 
-  if (!pageStr) return;
-
-  params.set("page", nextPage.toString());
   const movies = await getCurrentMovieList(nextPage, query);
 
   if (!movies) return;
   addMoviePost(movies.results, $movieList);
   disableMoreButton(movies.total_pages, currentPage, movies.results);
-
-  const newUrl = `${window.location.pathname}?${params.toString()}`;
-  history.pushState(null, "", newUrl);
 }
 
 async function getCurrentMovieList(page: number, query: string | null) {
